+++ src/main/java/com/soda/member/domain/AuthService.java
@@ -1,6 +1,6 @@
 package com.soda.member.domain;
 
-import com.soda.common.mail.service.EmailService;
+import com.soda.member.application.EmailService;
 import com.soda.global.response.GeneralException;
 import com.soda.global.security.jwt.JwtTokenProvider;
 import com.soda.member.domain.company.Company;
