+++ src/main/java/com/soda/member/infrastructure/company/CompanyProviderImpl.java
@@ -1,4 +1,4 @@
-package com.soda.member.infrastructure;
+package com.soda.member.infrastructure.company;
 
 import com.soda.member.domain.company.Company;
 import com.soda.member.domain.company.CompanyProvider;
